requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pypdf
Pillow
pdf2image
pytesseract
spacy
torch
transformers
sentence-transformers
faiss-cpu
langchain
email-validator
langchain-huggingface  
langchain-core        
langchain-community     
pandas